// src/app/models/admin/capphephoatdongkhoangsan/cptanthukhoangsan/cptanthuloaikhoangsan.model.ts
export class OutputCpTanThuLoaiKhoangSanModel {
  idtanthuloaikhoangsan: string;
  idloaikhoangsan: string;
  tenloaikhoangsan: string;
  idcappheptanthu: string;
  tenkhoangsan: string;
  siteid: string;
}

export class InputCpTanThuLoaiKhoangSanModel {
  idtanthuloaikhoangsan: string;
  idloaikhoangsan: string;
  idcappheptanthu: string;
  tenkhoangsan: string;
  siteid: string;
}